refactor: update notification services and setup logic
