+++ src/notifications/config/setupNotifications.ts
@@ -3,7 +3,6 @@ import * as Notifications from "expo-notifications";
 export const setupNotifications = () => {
   Notifications.setNotificationHandler({
     handleNotification: async () => ({
-      shouldShowAlert: true,
       shouldShowBanner: true,
       shouldShowList: false,
       shouldPlaySound: true,
