+++ src/notifications/index.ts
@@ -1,4 +1,3 @@
 export { usePushNotifications } from "./hooks/usePushNotifications";
-export { PushNotificationMessage } from "./types/push.types";
 export { PushNotificationViewModel } from "./viewModels/PushNotificationViewModel";
